# contracts/openapi/app-api.yaml
openapi: 3.1.0
info:
  title: Goat OS App API
  version: 0.1.0
  summary: Goat OS app and operating API for field tasks, goat lookup, protocol config, vaccination, and process integrity.
  description: REST/JSON API for operator/mobile and current admin-web operating surfaces. Reads and writes are RBAC
    scope-filtered before response DTOs are built.
servers:
  - url: https://api.goatos.example.invalid
    description: Synthetic contract server placeholder.
security:
  - bearerAuth: []
...
paths:
  /admin/tasks/review-fanouts/retry:
    post:
      tags:
        - SOP Admin
      operationId: retrySOPReviewFanouts
      summary: Retry pending or failed task review fanout rows.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetryReviewFanoutsRequest"
      responses:
        "200":
          description: Review fanouts retried.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetryReviewFanoutsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/ServerError"
...
  /app/tasks:
    get:
      tags:
        - Operator Tasks
      operationId: listAppTasks
      summary: List assigned operator tasks.
      parameters:
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Assigned tasks.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /app/tasks/{task_id}:
    get:
      tags:
        - Operator Tasks
      operationId: getAppTask
      summary: Get task detail and pinned SOP version.
      parameters:
        - $ref: "#/components/parameters/TaskId"
      responses:
        "200":
          description: Task detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "404":
          $ref: "#/components/responses/NotFoundOrNotAllowed"
  /app/sop-versions/{sop_version_id}:
    get:
      tags:
        - Operator Tasks
      operationId: getAppSOPVersion
      summary: Download a pinned SOP version.
      parameters:
        - $ref: "#/components/parameters/SOPVersionId"
      responses:
        "200":
          description: SOP version.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SOPVersionResponse"
        "404":
          $ref: "#/components/responses/NotFoundOrNotAllowed"
...
  /app/tasks/{task_id}/submissions:
    post:
      tags:
        - Operator Tasks
      operationId: submitAppTask
      summary: Submit an idempotent SOP task response.
      parameters:
        - $ref: "#/components/parameters/TaskId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitTaskRequest"
      responses:
        "200":
          description: Submission result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubmissionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/WriteConflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TaskId:
      name: task_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    SOPVersionId:
      name: sop_version_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Validation error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unauthorized:
      description: Missing or invalid authentication.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Permission denied.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotFoundOrNotAllowed:
      description: Resource is missing or outside the caller's visibility scope.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    WriteConflict:
      description: Optimistic concurrency, stale SOP version, or idempotency conflict.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    ValidationIssue:
      type: object
      additionalProperties: false
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
        code:
          type: string
        message:
          type: string
    ValidationReport:
      type: object
      additionalProperties: false
      required:
        - valid
        - errors
        - warnings
      properties:
        valid:
          type: boolean
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
        warnings:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
    SOPVersion:
      type: object
      additionalProperties: false
      required:
        - sop_version_id
        - tenant_id
        - sop_id
        - sop_code
        - version
        - version_label
        - status
        - form_dsl
        - proof_policy
        - compatibility
        - validation_report
        - row_version
        - created_at
        - updated_at
      properties:
        sop_version_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        sop_id:
          type: string
          format: uuid
        sop_code:
          type: string
        version:
          type: integer
          minimum: 1
        version_label:
          type: string
        status:
          type: string
          enum:
            - draft
            - published
            - retired
        form_dsl:
          type: object
          additionalProperties: true
        proof_policy:
          type: object
          additionalProperties: true
        compatibility:
          type: object
          additionalProperties: true
        validation_report:
          $ref: "#/components/schemas/ValidationReport"
        published_at:
          type:
            - string
            - "null"
          format: date-time
        retired_at:
          type:
            - string
            - "null"
          format: date-time
        row_version:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SOPVersionResponse:
      type: object
      additionalProperties: false
      required:
        - version
        - trace_id
      properties:
        version:
          $ref: "#/components/schemas/SOPVersion"
        trace_id:
          type: string
    TaskSummary:
      type: object
      additionalProperties: false
      required:
        - task_id
        - tenant_id
        - sop_id
        - sop_version_id
        - sop_code
        - task_type
        - title
        - description
        - state
        - scope_type
        - scope_id
        - priority
        - context
        - row_version
        - created_at
        - updated_at
      properties:
        task_id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        sop_id:
          type: string
          format: uuid
        sop_version_id:
          type: string
          format: uuid
        sop_code:
          type: string
        task_type:
          type: string
        title:
          type: string
        description:
          type: string
        state:
          type: string
          enum:
            - queued
            - assigned
            - in_progress
            - submitted
            - accepted
            - needs_review
            - rework_requested
            - rejected
            - canceled
        assigned_to:
          type:
            - string
            - "null"
          format: uuid
        scope_type:
          type: string
        scope_id:
          type: string
          format: uuid
        priority:
          type: string
        due_at:
          type:
            - string
            - "null"
          format: date-time
        context:
          type: object
          additionalProperties: true
        row_version:
          type: integer
          minimum: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    TaskListResponse:
      type: object
      additionalProperties: false
      required:
        - items
        - trace_id
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TaskSummary"
        trace_id:
          type: string
    ProofReference:
      type: object
      additionalProperties: false
      required:
        - proof_id
        - proof_type
        - subject_type
        - upload_state
        - metadata
      properties:
        proof_id:
          type: string
        proof_type:
          type: string
          enum:
            - photo
            - video
            - attachment
        subject_type:
          type: string
          enum:
            - batch
            - goat
            - shed
            - task
            - vial_lot
            - administration
            - other
        subject_id:
          type:
            - string
            - "null"
        upload_state:
          type: string
          enum:
            - pending
            - uploading
            - completed
            - failed
        metadata:
          type: object
          additionalProperties: true
...
    SubmissionItem:
      type: object
      additionalProperties: false
      required:
        - item_id
        - item_key
        - state
        - result
      properties:
        item_id:
          type: string
          format: uuid
        goat_id:
          type:
            - string
            - "null"
          format: uuid
        item_key:
          type: string
        state:
          type: string
          enum:
            - accepted
            - needs_review
            - rejected
            - skipped
        result:
          type: object
          additionalProperties: true
    SubmissionSummary:
      type: object
      additionalProperties: false
      required:
        - submission_id
        - task_id
        - sop_version_id
        - submitted_by
        - idempotency_key
        - answers
        - proof_refs
        - state
        - validation_report
        - items
        - submitted_at
        - row_version
      properties:
        submission_id:
          type: string
          format: uuid
        task_id:
          type: string
          format: uuid
        sop_version_id:
          type: string
          format: uuid
        submitted_by:
          type: string
          format: uuid
        idempotency_key:
          type: string
        answers:
          type: object
          additionalProperties: true
        proof_refs:
          type: array
          items:
            $ref: "#/components/schemas/ProofReference"
        state:
          type: string
          enum:
            - submitted
            - accepted
            - needs_review
            - rejected
            - voided
        validation_report:
          $ref: "#/components/schemas/ValidationReport"
        items:
          type: array
          items:
            $ref: "#/components/schemas/SubmissionItem"
        submitted_at:
          type: string
          format: date-time
        accepted_at:
          type:
            - string
            - "null"
          format: date-time
        row_version:
          type: integer
          minimum: 1
    TaskResponse:
      type: object
      additionalProperties: false
      required:
        - task
        - trace_id
      properties:
        task:
          $ref: "#/components/schemas/TaskSummary"
        sop_version:
          oneOf:
            - $ref: "#/components/schemas/SOPVersion"
            - type: "null"
        submissions:
          type: array
          items:
            $ref: "#/components/schemas/SubmissionSummary"
        trace_id:
          type: string
    RetryReviewFanoutsRequest:
      type: object
      additionalProperties: false
      properties:
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 50
    RetryReviewFanoutsResponse:
      type: object
      additionalProperties: false
      required:
        - applied
        - trace_id
      properties:
        applied:
          type: integer
          minimum: 0
        trace_id:
          type: string
    SubmitTaskRequest:
      type: object
      additionalProperties: false
      required:
        - sop_version_id
        - idempotency_key
        - answers
        - proof_refs
      properties:
        sop_version_id:
          type: string
          format: uuid
        idempotency_key:
          type: string
          minLength: 8
          maxLength: 200
        answers:
          type: object
          additionalProperties: true
        proof_refs:
          type: array
          items:
            $ref: "#/components/schemas/ProofReference"
    SubmissionResponse:
      type: object
      additionalProperties: false
      required:
        - submission
        - task
        - trace_id
      properties:
        submission:
          $ref: "#/components/schemas/SubmissionSummary"
        task:
          $ref: "#/components/schemas/TaskSummary"
        trace_id:
          type: string
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - code
        - message
        - field_errors
        - trace_id
        - retryable
      properties:
        code:
          type: string
          minLength: 1
          maxLength: 120
        message:
          type: string
          minLength: 1
          maxLength: 500
        field_errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        trace_id:
          type: string
          minLength: 1
          maxLength: 200
        retryable:
          type: boolean
    FieldError:
      type: object
      additionalProperties: false
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
          minLength: 1
          maxLength: 200
        code:
          type: string
          minLength: 1
          maxLength: 120
        message:
          type: string
          minLength: 1
          maxLength: 500
...